pyproject.toml
[project]
name = "multi-agent-solver-direct"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.2",
    "langfuse>=3.8.1",
    "langgraph>=1.0.1",
    "openai>=2.6.1",
    "pillow>=12.0.0",
    "rich>=14.2.0",
]
